Add files via upload
This files is Python file to run and control the instrument. This file has a function as data acquisition program to obtain data from sensors. The sensors read some physical phenomena in sample such as conductivity, resistance, capacitance, Total solid dissolve material, and temperature. All of data send those value to the raspberry pi using SERIAL communication conducted by Microcontroller Arduino Uno R3. The program then send the data to MYSQL database. The program also displays the data to the monitor of the instruments. The program made in Python IDE 3.6